Update install.pp
This change is to provide support for windows server core. I ran into issues with the exec unzip resource (wouldnt extract zip file) when applying to windows core servers, puppet did not pick up the error for some reason so when i tried to run the shell com application .net objects individually on a core server it gave me the resulted failure...
Error HRESULT E_FAIL has been returned from a call to a COM component.
At line:1 char:39
+ ... Application;$sh.namespace((Convert-Path "$install_dir")).Copyhere($sh ...
+                 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : OperationStopped: (:) [], COMException
    + FullyQualifiedErrorId : System.Runtime.InteropServices.COMException

So instead, in this commit, we should leverage the archive resource for downloading, unzipping and cleaning, and as per the concerns about 7zip, puppetforge says that we can leverage archive for zip files without a problem, as long as servers are at a minimum 2012. (otherwise if other type of archive, 7zip would be required). ref: https://forge.puppet.com/puppet/archive#setup